Add code and experiments for Logistic regression classifier
I copied and slightly modified the code realted to Logistic regression
and implementation of a Logistic regression classifier on the base of
Adaline implementation for Chapter 2. I added some explanations about
the essence of the cost function based on the logistic sigmoid function
and the main differences between this classifier and the Adaline one.
I also added experiments and checked the difference between 2 different
multi-class classification techniques.